Fix bug #57
When judging the valid variable names in runtime, can not use
`!isdigit`, since the first character of negative number is `-`, which
is non digital. Changed to the way followed the grammar rule
definition, e.g `isalpha` for user defined variable and starting with
underscore `var_one[0] == ‘_’` for inner temp variables.