Fix bug #62
In the byte code, while dealing with the `for` loop, open a new scope
for the stmt_list, so the variable defined inside the stmt_list scope
won’t effect on the outer `for` scope loop variable. After exit the
stmt_list, using `stmt_list_end` to close the scope, and back to loop
scope.